Local development changes

+++ docker/deploy_contracts.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
 NODEOS_RUNNING=$1
+RUNNING_IN_GITPOD=$2
 
 set -m
 
@@ -8,13 +9,28 @@ set -m
 SYSTEM_ACCOUNT_PRIVATE_KEY="5KQwrPbwdL6PhXujxW37FSSQZ1JiwsST4cqQzDeyXtP79zkvFD3"
 SYSTEM_ACCOUNT_PUBLIC_KEY="EOS6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5GDW5CV"
 
-ROOT_DIR="/home/gitpod"
-CONTRACTS_DIR="$ROOT_DIR/contracts"
-BLOCKCHAIN_DATA_DIR=$ROOT_DIR/eosio/chain/data
-BLOCKCHAIN_CONFIG_DIR=$ROOT_DIR/eosio/chain/config
-WALLET_DIR="$ROOT_DIR/eosio-wallet"
+if [ -z "$RUNNING_IN_GITPOD" ]; then
+  echo "Running locally..."
+  ROOT_DIR="/opt"
+  CONTRACTS_DIR="$ROOT_DIR/eosio/bin/contracts"
+  BLOCKCHAIN_DATA_DIR=/root/.local/share
+  BLOCKCHAIN_CONFIG_DIR=/opt/eosio/bin/config-dir
+  WALLET_DIR="/root/eosio-wallet"
+else
+  echo "Running in Gitpod..."
+  export NODEOS_URL=$(gp url 8888)
+  ROOT_DIR="/home/gitpod"
+  CONTRACTS_DIR="/opt/eosio/bin/contracts"
+  BLOCKCHAIN_DATA_DIR=$ROOT_DIR/eosio/chain/data
+  BLOCKCHAIN_CONFIG_DIR=$ROOT_DIR/eosio/chain/config
+  WALLET_DIR="$ROOT_DIR/eosio-wallet"
+fi
 
-CONFIG_DIR="$ROOT_DIR/config-dir"
+mkdir -p $ROOT_DIR/bin
+
+PATH="$PATH:$ROOT_DIR/bin:$ROOT_DIR/bin/scripts"
+GITPOD_WORKSPACE_ROOT="/workspace/sample-return-values-application"
+CONFIG_DIR="$ROOT_DIR/bin/config-dir"
 
 function post_preactivate {
   curl -X POST http://127.0.0.1:8888/v1/producer/schedule_protocol_feature_activations -d '{"protocol_features_to_activate": ["0ec7e080177b2c02b278d5088611686b49d739925a92d9bfcacd7fc6b74053bd"]}'
@@ -76,10 +92,22 @@ function setabi {
 }
 
 # Move into the executable directory
-cd $ROOT_DIR/
+cd $ROOT_DIR/bin/
 mkdir -p $CONFIG_DIR
 mkdir -p $BLOCKCHAIN_DATA_DIR
 mkdir -p $BLOCKCHAIN_CONFIG_DIR
+
+function start_wallet {
+  echo "Starting the wallet"
+  rm -rf $WALLET_DIR
+  mkdir -p $WALLET_DIR
+  nohup keosd --unlock-timeout 999999999 --wallet-dir $WALLET_DIR --http-server-address 127.0.0.1:8900 2>&1 &
+  sleep 1s
+  wallet_password=$(cleos wallet create --to-console | awk 'FNR > 3 { print $1 }' | tr -d '"')
+  echo $wallet_password > "$CONFIG_DIR"/keys/default_wallet_password.txt
+
+  cleos wallet import --private-key $SYSTEM_ACCOUNT_PRIVATE_KEY
+}
 
 if [ -z "$NODEOS_RUNNING" ]; then
   echo "Starting the chain for setup"
@@ -113,7 +141,7 @@ sleep 2s
 echo "Creating accounts and deploying contracts"
 
 sleep 1s
-cleos wallet unlock --password </password
+start_wallet
 cleos create account eosio returnvalue $SYSTEM_ACCOUNT_PUBLIC_KEY
 
 # preactivate concensus upgrades
@@ -151,9 +179,12 @@ cleos set code returnvalue $CONTRACTS_DIR/action_results/action_results.wasm -p 
 
 echo "All done initializing the blockchain"
 
-if [[ -z $NODEOS_RUNNING ]]; then
-  echo "Shut down Nodeos, sleeping for 2 seconds to allow time for at least 4 blocks to be created after deploying contracts"
-  sleep 2s
-  kill %1
-  fg %1
+# If running in Gitpod, we *don't* want to shutdown the blockchain; we'll leave it running in the terminal window.
+if [ -z "$RUNNING_IN_GITPOD" ]; then
+  if [[ -z $NODEOS_RUNNING ]]; then
+    echo "Shut down Nodeos, sleeping for 2 seconds to allow time for at least 4 blocks to be created after deploying contracts"
+    sleep 2s
+    kill %1
+    fg %1
+  fi
 fi
